Narrow mdlManager 0031b680 integer parameter

+++ src/Event/Fcl/fclShopMisc.c
@@ -275,7 +275,7 @@ void FUN_003f3fc0(int param_1,int param_2,u32 param_3,u64 param_4);
 void FUN_003f4350(int param_1,int param_2,int param_3,int param_4);
 void FUN_003f45e0(int param_1,int param_2,int param_3,int param_4);
 void FUN_003f4870(int param_1,int param_2,int param_3,int param_4);
-u64 FUN_003f4a90(int param_1,int param_2,int param_3,int param_4,int param_5,short *param_6);
+u64 FUN_003f4a90(int param_1,int param_2,int param_3,int param_4,u64 param_5,short *param_6);
 void FUN_003f4de0(int param_1,int param_2,u32 param_3,int param_4);
 void FUN_003f55b0(u64 param_1,u64 param_2,u32 param_3,int param_4,int param_5, u32 param_6);
 #pragma alias FUN_003f55b0_i FUN_003f55b0
@@ -4464,7 +4464,7 @@ void FUN_003f4870(int param_1,int param_2,int param_3,int param_4)
 
 u64
 
-FUN_003f4a90(int param_1,int param_2,int param_3,int param_4,int param_5,short *param_6)
+FUN_003f4a90(int param_1,int param_2,int param_3,int param_4,u64 param_5,short *param_6)
